requirements.txt
streamlit
pandas
matplotlib
python-docx
num2words
openpyxl
